+++ inst/GerminaQuant/pkgs.R
@@ -4,19 +4,20 @@
 #> open https://flavjack.github.io/GerminaR/
 #> open https://flavjack.shinyapps.io/germinaquant/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2020-04-20
+#> date .: 2020-04-28
 # -------------------------------------------------------------------------
 
-suppressPackageStartupMessages({
-  library(GerminaR)
-  library(shiny)
-  library(metathis)
-  library(tidyverse)
-  library(shinydashboard)
-  library(shinyWidgets)
-  library(gsheet)
-  library(readxl)
-  library(cowplot)
-})
+# -------------------------------------------------------------------------
+# packages ----------------------------------------------------------------
+# -------------------------------------------------------------------------
 
+library(GerminaR)
+library(shiny)
+library(metathis)
+library(tidyverse)
+library(shinydashboard)
+library(shinyWidgets)
+library(gsheet)
+library(readxl)
+library(cowplot)