+++ NEWS.md
@@ -4,13 +4,17 @@
   - `fplot()` 
     - Absolute values to ybrakes
     - Include color using vectors
+    - Font size in significance
+    - Delete plot borders
   - Added function `textcolor()`
+  - Update logos
   
 - *GerminaQuant*
   - Styled messages in the console
-  - Include multiple layers in the graphs
+  - Include multiple layers in the graphs `opt`
   - Possibility to exclude the error bar in the plots
   - Include demo videos for the app
+  - Change name: Model plot --> Model diagnostic
   
 # Version 2.1.1
 
@@ -69,6 +73,6 @@ Changes incompatible with the old versions. Arguments changed in the syntax for 
 - Fix resolution in the graphics produces in GerminaQuant
 - Auto close local session of app when finish process
 
-# Version 1.0
+# Version 1.0.0
 
 - Package release